requirements.txt
numpy==1.16.4
jieba==0.39
grpcio>=1.19.0
Flask==1.1.1
requests>=2.21.0
scikit-learn>=0.21.1
retrying
pymilvus==1.0.1
pandas==0.24.2
six>=1.15.0
setuptools>=41.0.0
gunicorn==20.0.4
eventlet==0.28.0
psycopg2_binary==2.8.6
elasticsearch==7.10.0
redlock>=1.2.0
celery==4.4.0
redis==3.5.3
pydantic==1.8.1
raven==6.10.0
flask-redis==0.4.0
Flask-SQLAlchemy==2.4.1
Flask-Mail==0.9.1